👌(layout) changed default header level to 2 for "section_cadenced"
It has been reported HTML validator warnings about some multiple "h1"
elements on homepage. This was due to "section_cadenced.html" template
from SectionPlugin that had default header level set to 1.

We set default header level to 2 so it won't never create "h1" from common
context. "h1" has a special semantic meaning and so it should never be
something as a default behavior from a section plugin. Still, there is a
"h1" on homepage from introduction with LargeBannerPlugin template which
have a default level title to 1. This is ok since LargeBannerPlugin has been
done for introduction which is best place for a "h1", but we have to remember
this for futur usage.