📦(frontend) group all frontend dependencies under "dependencies"
Dependencies were split between `dependencies` and `devDependencies`.
This was mostly an aesthetic issue as functionally the package needs
all of them to build and we never use `yarn install --production`
since we have a frontend bundle that needs building, not a backend
app that needs to be deployed and pull its dependencies.

It caused issues in that consumers of our published NPM package (whose
main function is to enable users to build their own version of richie)
were unable to do so due to missing the `devDependencies`. Moving
everything under `dependencies` means users now have everything they
need to build their own richie frotend.