♿️(search) fix frontend pagination accessibility issue
There was an issue with the HTML tags used for pagination. For some
reason, <ul> is an invalid tag for role="navigation". We just added
an empty <nav> around our <ul> navigation with the aria-labelledby
pointing to it.

This solves an issue screen readers would have encountered seeing the
pagination as a navigational element.